+++ packages/database/src/migrations/0023_draft_undone_at.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "document_yjs_drafts" ADD COLUMN "undone_at" timestamp with time zone;